Time: 0 ms (100%), Space: 8.1 MB (5.98%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -3,7 +3,7 @@ public:
     long mySqrt(int x) {
         int i=0 ,j=x;
         long ans=0;
-        long mid;
+        long long mid;
         while(i<=j){
             mid=i+(j-i)/2;
             if(mid*mid==x)
